Avoid showing the 130 error code after hitting ^C
When a user hits ^C on command line to cancel entering a lengthy command,
$! in __lp_set_prompt is set to 130 (128 + SIGINT). Displaying this error
on the input line (especially in warning colours) makes little sense.
Therefore, liquidprompt should check if the command submitted is a real
command or is it part of PROMPT_COMMAND in which case the error does
not come from a user's command and should not be presented.